lots of renaming

+++ polaris-pipeline/coordinator/Clients/TextExtractor/ITextExtractorClient.cs
@@ -9,10 +9,10 @@ namespace coordinator.Clients.TextExtractor
 {
         public interface ITextExtractorClient
         {
-                Task<StoreCaseIndexesResult> StoreCaseIndexesAsync(string documentId, string urn, long caseId, long versionId, string blobName, Guid correlationId, Stream ocrResults);
-                Task<IList<StreamlinedSearchLine>> SearchTextAsync(string urn, long caseId, string searchTerm, Guid correlationId);
-                Task<IndexDocumentsDeletedResult> RemoveCaseIndexesAsync(string urn, long caseId, Guid correlationId);
-                Task<SearchIndexCountResult> GetCaseIndexCount(string urn, long caseId, Guid correlationId);
-                Task<SearchIndexCountResult> GetDocumentIndexCount(string urn, long caseId, string documentId, long versionId, Guid correlationId);
+                Task<StoreCaseIndexesResult> StoreCaseIndexesAsync(string documentId, string urn, int caseId, long versionId, string blobName, Guid correlationId, Stream ocrResults);
+                Task<IList<StreamlinedSearchLine>> SearchTextAsync(string urn, int caseId, string searchTerm, Guid correlationId);
+                Task<IndexDocumentsDeletedResult> RemoveCaseIndexesAsync(string urn, int caseId, Guid correlationId);
+                Task<SearchIndexCountResult> GetCaseIndexCount(string urn, int caseId, Guid correlationId);
+                Task<SearchIndexCountResult> GetDocumentIndexCount(string urn, int caseId, string documentId, long versionId, Guid correlationId);
         }
 }
