+++ polaris-pipeline/text-extractor/Factories/SearchLineFactory.cs
@@ -8,7 +8,7 @@ namespace text_extractor.Factories;
 
 public class SearchLineFactory : ISearchLineFactory
 {
-    public SearchLine Create(long caseId, string documentId, long versionId, string blobName, ReadResult readResult, Common.Domain.SearchIndex.Line line, int index)
+    public SearchLine Create(int caseId, string documentId, long versionId, string blobName, ReadResult readResult, Common.Domain.SearchIndex.Line line, int index)
     {
         var id = $"{caseId}:{documentId}:{readResult.Page}:{index}";
         var bytes = Encoding.UTF8.GetBytes(id);
