+++ polaris-pipeline/text-extractor/Services/SearchIndexService/ISearchIndexService.cs
@@ -9,14 +9,14 @@ namespace text_extractor.Services.CaseSearchService
 {
     public interface ISearchIndexService
     {
-        Task<int> SendStoreResultsAsync(AnalyzeResults analyzeResults, string documentId, long caseId, long versionId, string blobName, Guid correlationId);
+        Task<int> SendStoreResultsAsync(AnalyzeResults analyzeResults, string documentId, int caseId, long versionId, string blobName, Guid correlationId);
 
-        Task<IList<StreamlinedSearchLine>> QueryAsync(long caseId, string searchTerm);
+        Task<IList<StreamlinedSearchLine>> QueryAsync(int caseId, string searchTerm);
 
-        Task<IndexDocumentsDeletedResult> RemoveCaseIndexEntriesAsync(long caseId, Guid correlationId);
+        Task<IndexDocumentsDeletedResult> RemoveCaseIndexEntriesAsync(int caseId, Guid correlationId);
 
-        Task<SearchIndexCountResult> GetCaseIndexCount(long caseId, Guid correlationId);
+        Task<SearchIndexCountResult> GetCaseIndexCount(int caseId, Guid correlationId);
 
-        Task<SearchIndexCountResult> GetDocumentIndexCount(long caseId, string documentId, long versionId, Guid correlationId);
+        Task<SearchIndexCountResult> GetDocumentIndexCount(int caseId, string documentId, long versionId, Guid correlationId);
     }
 }
